pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "python-dateutil>=2.9.0.post0",
    "matplotlib>=3.10.1",
    "openai>=1.77.0",
    "trafilatura>=2.0.0",
    "flask-wtf>=1.2.2",
    "oauthlib>=3.2.2",
]
